requirements.txt
python-aalib
pypot
flask
gcloud
sseclient
requests_toolbelt
python_jwt
pycrypto
prettytable
flask_cors
opencv-python-headless
mediapipe
tensorflow
numpy
picamera
Pillow
